Refactor database library
